Add functions and delet superfluous

+++ libftprintf/ft_putnbr_fd.c
@@ -12,7 +12,7 @@ static void	print_symbole(long int n, int fd)
 	ft_putchar_fd((n % 10 + '0'), fd);
 }
 
-void		ft_putnbr_fd(int n, int fd)
+void		ft_putnbr_fd(long int n, int fd)
 {
 	if (n < 0)
 	{
